+++ Tests/SentryTests/SentryTests-Bridging-Header.h
@@ -81,4 +81,5 @@
 #import "SentryTracer.h"
 #import "SentryTransport.h"
 #import "SentryTransportFactory.h"
+#import "SentryUIViewControllerSanitizer.h"
 #import "SentryUserFeedback.h"